requirements.txt
flask==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9

pymupdf==1.24.9
pillow==10.4.0
pytesseract==0.3.10

langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-openai==0.1.22
langchain-community==0.2.16
